Rename get_data.py to income.py
