--- esam_EvolutionDesSalairesAnnuelsMoyens/README.md
@@ -1,21 +1,0 @@
-# PYBD
-richard.lay et steven.tien
-
-# L'évolution du salaire annuel moyen dans le monde
-## But
-Dans ce projet montrons l'évolution du salaire moyen ainsi que de son influence.
-## Prérequis
-Voir requirement.txt
-## Le projet
-Pour lancer le projet juste lancer le fichier python main.py
-    python3 main.py
-
-
-
-
-Sources :
-PIB : https://data.oecd.org/gdp/gross-domestic-product-gdp.htm
-IDH : https://hdr.undp.org/en/indicators/137506#
-PaysIso : https://gist.github.com/radcliff/f09c0f88344a7fcef373
-Salaire Moyen dans le monde : https://stats.oecd.org/viewhtml.aspx?datasetcode=AV_AN_WAGE&lang=fr
-Inégalité homme/femme : https://ec.europa.eu/eurostat/databrowser/view/SDG_05_20/default/table?lang=en
