--- esam_EvolutionDesSalairesAnnuelsMoyens/requirement.txt
@@ -1,16 +1,0 @@
-numpy
-xarray
-pandas
-pandas-datareader
-beautifulsoup4
-lxml
-openpyxl
-xlrd
-imageio
-folium
-seaborn
-plotly
-dash
-gunicorn
-flask
-scipy
