+++ strl_EvolutionDesSalairesAnnuelsMoyens/data/sdg_05_20_tabular-1.tsv
@@ -1,0 +1,41 @@
+freq,unit,nace_r2,geo\TIME_PERIOD	2002 	2006 	2007 	2008 	2009 	2010 	2011 	2012 	2013 	2014 	2015 	2016 	2017 	2018 	2019 	2020 
+A,PC,B-S_X_O,AL	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	6.8 	: 	: 
+A,PC,B-S_X_O,AT	: 	25.5 	25.5 	25.1 	24.3 	24.0 	23.5 	22.9 	22.3 	22.2 	21.8 	20.8 	20.7 	20.4 	19.9 p	18.9 p
+A,PC,B-S_X_O,BE	: 	9.5 	10.1 	10.2 	10.1 	10.2 	9.4 	8.3 	7.5 	6.6 	6.4 	6.0 	5.8 	5.8 	5.8 p	5.3 p
+A,PC,B-S_X_O,BG	18.9 	12.4 	12.1 	12.3 	13.3 	13.0 	13.2 	15.1 	14.1 	14.2 	15.5 	14.6 	14.3 	13.9 	14.1 p	12.7 p
+A,PC,B-S_X_O,CH	: 	18.6 	: 	18.4 d	18.4 d	17.8 	17.6 	17.4 	17.6 	17.4 	17.9 	17.4 	17.6 	18.6 	18.6 p	18.4 p
+A,PC,B-S_X_O,CY	22.5 	21.8 	22.0 	19.5 	17.8 	16.8 	16.1 	15.6 	14.9 	14.2 	13.2 	12.3 	11.2 	10.4 	10.1 p	9.0 p
+A,PC,B-S_X_O,CZ	22.1 	23.4 	23.6 	26.2 	25.9 	21.6 	22.6 	22.5 	22.3 	22.5 	22.5 d	21.5 d	21.1 d	20.1 d	19.2 d	16.4 d
+A,PC,B-S_X_O,DE	: 	22.7 	22.8 	22.8 	22.6 	22.3 	22.4 	22.7 	22.1 	22.3 	21.8 	21.1 	20.4 	20.1 	19.2 p	18.3 p
+A,PC,B-S_X_O,DK	: 	17.6 	17.7 	17.1 	16.8 	17.1 	16.4 	16.8 	16.5 	16.0 	15.1 	15.1 	14.8 	14.6 	14.0 	13.9 
+A,PC,B-S_X_O,EA19	: 	: 	: 	: 	: 	17.0 	17.3 	17.6 	17.1 	16.8 	16.5 	16.2 	15.7 	15.3 	14.6 p	14.1 p
+A,PC,B-S_X_O,EE	: 	29.8 	30.9 	27.6 	26.6 	27.7 	27.3 	29.9 	29.8 	28.1 	26.7 	24.8 	24.9 	21.8 	21.7 p	21.1 p
+A,PC,B-S_X_O,EL	25.5 	20.7 	21.5 	22 d	: 	15.0 	: 	: 	: 	12.5 	: 	: 	: 	10.4 	: 	: 
+A,PC,B-S_X_O,ES	20.2 	17.9 	18.1 	16.1 	16.7 	16.2 	17.6 	18.7 	17.8 	14.9 	14.1 	14.8 	13.5 	11.9 	9.4 	9.4 p
+A,PC,B-S_X_O,EU27_2007	: 	17.7 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 	: 
+A,PC,B-S_X_O,EU27_2020	: 	: 	: 	: 	: 	15.8 	16.2 	16.4 	16.0 	15.7 	15.5 	15.1 	14.6 	14.4 	13.7 p	13.0 p
+A,PC,B-S_X_O,EU28	: 	: 	: 	: 	: 	17.1 	17.1 	17.4 	16.8 	16.6 	16.4 	16.1 	15.7 	15.3 	: 	: 
+A,PC,B-S_X_O,FI	: 	21.3 	20.2 	20.5 	20.8 	20.3 	19.1 	19.2 	18.8 	18.4 	17.5 	17.5 	17.1 	16.9 	16.6 	16.7 
+A,PC,B-S_X_O,FR	: 	15.4 	17.3 	16.9 	15.2 	15.6 	15.7 	15.6 	15.5 	15.5 	15.6 	15.9 	16.3 	16.7 	16.2 p	15.8 p
+A,PC,B-S_X_O,HR	: 	: 	: 	: 	: 	5.7 	: 	: 	7.7 	8.7 	: 	11.6 	12.3 	11.4 	11.5 p	11.2 p
+A,PC,B-S_X_O,HU	19.1 	14.4 	16.3 	17.5 	17.1 	17.6 	18.0 	20.1 	18.4 	15.1 	14.0 	14.0 	15.9 b	14.2 	18.2 b	17.2 
+A,PC,B-S_X_O,IE	15.1 	17.2 	17.3 	12.6 	12.6 	13.9 	12.7 	12.2 	12.9 	13.9 	13.9 p	14.2 p	14.4 p	11.3 	: 	: 
+A,PC,B-S_X_O,IS	: 	: 	24.0 	20.7 	18.0 	17.7 	17.8 	17.7 	19.0 	16.7 	17.5 	15.8 	15.3 	13.8 	14.2 p	13.0 p
+A,PC,B-S_X_O,IT	: 	4.4 	5.1 	4.9 	5.5 	5.3 	5.7 	6.5 	7.0 	6.1 	5.5 p	5.3 p	5 p	5.5 	4.7 p	4.2 p
+A,PC,B-S_X_O,LT	13.2 	17.1 	22.6 	21.6 	15.3 	11.9 b	11.5 	11.9 	12.2 	13.3 	14.2 	14.4 	15.2 	14.0 	13.3 p	13.0 p
+A,PC,B-S_X_O,LU	: 	10.7 	10.2 	9.7 	9.2 	8.7 	7.9 	7.0 	6.2 	5.4 	4.7 	3.9 	2.6 	1.4 	1.3 p	0.7 p
+A,PC,B-S_X_O,LV	: 	15.1 	13.6 	11.8 	13.1 	15.5 	14.1 	14.9 	16.0 	17.3 	18.4 	19.7 	19.8 	19.6 	21.2 p	22.3 p
+A,PC,B-S_X_O,ME	: 	: 	: 	: 	: 	: 	: 	: 	: 	7.7 	: 	: 	: 	: 	: 	: 
+A,PC,B-S_X_O,MK	: 	: 	: 	: 	: 	: 	: 	: 	: 	9.1 	: 	: 	: 	: 	: 	: 
+A,PC,B-S_X_O,MT	: 	5.2 	7.8 	9.2 	7.7 	7.2 	7.7 	9.5 	9.7 	10.6 	10.7 	11.6 	13.2 	13.0 	11.6 p	10.0 p
+A,PC,B-S_X_O,NL	18.7 	23.6 	19.3 	18.9 	18.5 	17.8 	18.8 	18.0 	17.2 	17.0 	16.1 	15.6 	15.1 	14.7 	14.6 p	14.2 p
+A,PC,B-S_X_O,NO	: 	16.0 	15.6 	17.0 	16.5 	16.1 	15.7 	14.7 	15.5 	14.5 	16.0 	14.5 	13.7 	13.2 	13.2 dp	13.4 dp
+A,PC,B-S_X_O,PL	7.5 	7.5 	14.9 	11.4 	8.0 	4.5 	5.5 	6.4 	7.1 	7.7 	7.3 	7.1 	7.0 	8.5 	6.5 	4.5 
+A,PC,B-S_X_O,PT	: 	8.4 	8.5 	9.2 	10.0 	12.8 	12.9 	15.0 	13.3 	14.9 	16.0 	13.9 	10.8 	8.9 	10.9 p	11.4 p
+A,PC,B-S_X_O,RO	16.0 	7.8 	12.5 d	8.5 d	7.4 d	8.8 	9.6 	6.9 	4.9 	4.5 	5.6 e	4.8 e	2.9 e	2.2 	3.3 e	2.4 e
+A,PC,B-S_X_O,RS	: 	: 	: 	: 	: 	: 	: 	: 	: 	8.7 	: 	: 	: 	9.6 	: 	: 
+A,PC,B-S_X_O,SE	: 	16.5 	17.8 	16.9 	15.7 	15.4 	15.6 	15.5 	14.6 	13.8 	14.0 	13.3 	12.5 	12.1 	11.8 	11.2 
+A,PC,B-S_X_O,SI	6.1 	8.0 	5.0 	4.1 	-0.9 	0.9 	3.3 	4.5 	6.3 	7.0 	8.2 	8.1 	8.4 	9.3 	7.9 p	3.1 p
+A,PC,B-S_X_O,SK	27.7 	25.8 	23.6 	20.9 	21.9 	19.6 	20.1 	20.8 	18.8 	19.7 	19.7 	19.2 	20.1 	19.8 	18.4 	15.8 
+A,PC,B-S_X_O,TR	: 	-2.2 	: 	: 	: 	: 	: 	: 	: 	-1.3 	: 	: 	: 	: 	: 	: 
+A,PC,B-S_X_O,UK	27.3 	24.3 	20.8 	21.4 	20.6 	23.3 b	21.8 	22.6 	21.0 	20.9 	21.0 	20.7 	20.8 	19.8 	: 	:
