+++ strl_EvolutionDesSalairesAnnuelsMoyens/requirement.txt
@@ -1,0 +1,16 @@
+numpy
+xarray
+pandas
+pandas-datareader
+beautifulsoup4
+lxml
+openpyxl
+xlrd
+imageio
+folium
+seaborn
+plotly
+dash
+gunicorn
+flask
+scipy
